Move the pool acquisition metric logic to the PgRepositoryFactory
